lab4 发送fin 发生ip从144.1 编程144.9

+++ libsponge/tcp_connection.hh
@@ -30,7 +30,7 @@ class TCPConnection {
     bool _isFinAckReceived{false};
 
     TCPState::State _tcpState{TCPState::State::LISTEN};
-    uint64_t _lastAckNo {0};
+    uint32_t _lastAckNo {0};
     enum class Type{
         SYN,
         FIN,
